Consistent size for socketID

+++ src/Sodaq_R4X.h
@@ -212,7 +212,7 @@ public:
     * Sockets
     *****************************************************************************/
 
-    int    socketCreate(uint16_t localPort = 0, Protocols protocol = UDP);
+    int8_t    socketCreate(uint16_t localPort = 0, Protocols protocol = UDP);
 
     bool   socketSetR4KeepAlive(uint8_t socketID);
     bool   socketSetR4Option(uint8_t socketID, uint16_t level, uint16_t optName, uint32_t optValue, uint32_t optValue2 = 0);
